requirements.txt
# Career AI Assistant — no chromadb, no protobuf, no conflicts

streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.9.0
pypdf>=4.0.0
PyGithub>=1.60.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
sentence-transformers>=3.0.0
networkx>=3.0
pydantic>=2.0.0

# Optional — better PDF extraction (install if you have them)
# PyMuPDF>=1.23.0
# pdfplumber>=0.10.0